feat: :sparkles: Add feature to customize Scylla CDC start time for Kafka connector
Introduced a new feature that allows customization of the start time for the Scylla CDC Kafka connector. Previously, the connector always began reading from the start of the TTL. With this update, users can now specify a custom start time, such as reading from the last 60 seconds. This improves flexibility for real-time streaming scenarios.